Remove old storage; Refactor FS structure
