+++ neuro_flow/cli/images.py
@@ -6,8 +6,8 @@ from neuro_flow.cli.click_types import LIVE_IMAGE_OR_ALL
 from neuro_flow.cli.root import Root
 from neuro_flow.cli.utils import argument, option, wrap_async
 from neuro_flow.live_runner import LiveRunner
-from neuro_flow.storage_api import ApiStorage
-from neuro_flow.storage_base import Storage2
+from neuro_flow.storage.api import ApiStorage
+from neuro_flow.storage.base import Storage
 
 
 if sys.version_info >= (3, 7):
@@ -33,7 +33,7 @@ async def build(root: Root, force_overwrite: bool, image: str) -> None:
     """
     async with AsyncExitStack() as stack:
         client = await stack.enter_async_context(neuro_sdk.get())
-        storage: Storage2 = await stack.enter_async_context(ApiStorage(client))
+        storage: Storage = await stack.enter_async_context(ApiStorage(client))
         runner = await stack.enter_async_context(
             LiveRunner(root.config_dir, root.console, client, storage, root)
         )
