+++ neuro_flow/cli/storage.py
@@ -6,8 +6,8 @@ from neuro_flow.cli.click_types import LIVE_VOLUME_OR_ALL
 from neuro_flow.cli.root import Root
 from neuro_flow.cli.utils import argument, wrap_async
 from neuro_flow.live_runner import LiveRunner
-from neuro_flow.storage_api import ApiStorage
-from neuro_flow.storage_base import Storage2
+from neuro_flow.storage.api import ApiStorage
+from neuro_flow.storage.base import Storage
 
 
 if sys.version_info >= (3, 7):
@@ -29,7 +29,7 @@ async def upload(
     use `upload ALL` for uploading all volumes."""
     async with AsyncExitStack() as stack:
         client = await stack.enter_async_context(neuro_sdk.get())
-        storage: Storage2 = await stack.enter_async_context(ApiStorage(client))
+        storage: Storage = await stack.enter_async_context(ApiStorage(client))
         runner = await stack.enter_async_context(
             LiveRunner(root.config_dir, root.console, client, storage, root)
         )
@@ -52,7 +52,7 @@ async def download(
     use `download ALL` for downloading all volumes."""
     async with AsyncExitStack() as stack:
         client = await stack.enter_async_context(neuro_sdk.get())
-        storage: Storage2 = await stack.enter_async_context(ApiStorage(client))
+        storage: Storage = await stack.enter_async_context(ApiStorage(client))
         runner = await stack.enter_async_context(
             LiveRunner(root.config_dir, root.console, client, storage, root)
         )
@@ -75,7 +75,7 @@ async def clean(
     use `clean ALL` for cleaning up all volumes."""
     async with AsyncExitStack() as stack:
         client = await stack.enter_async_context(neuro_sdk.get())
-        storage: Storage2 = await stack.enter_async_context(ApiStorage(client))
+        storage: Storage = await stack.enter_async_context(ApiStorage(client))
         runner = await stack.enter_async_context(
             LiveRunner(root.config_dir, root.console, client, storage, root)
         )
@@ -93,7 +93,7 @@ async def mkvolumes(
     """Create all remote folders for volumes."""
     async with AsyncExitStack() as stack:
         client = await stack.enter_async_context(neuro_sdk.get())
-        storage: Storage2 = await stack.enter_async_context(ApiStorage(client))
+        storage: Storage = await stack.enter_async_context(ApiStorage(client))
         runner = await stack.enter_async_context(
             LiveRunner(root.config_dir, root.console, client, storage, root)
         )
